Add files via upload
Title: The momentum of News; Author: Ying Wang, Bohui Zhang, Xiaoneng Zhu; Description: Relying on a comprehensive data set of news releases, we construct monthly firm-level news sentiment scores during the 2000–2016 period and document a news momentum phenomenon of stocks with more positive news in the past generating more positive news in the future. We propose three hypotheses to explain this phenomenon and find that news momentum is driven by the persistence of firms fundamentals rather than stale news or firms strategic disclosure. A trading strategy that combines a long position in a good news quintile portfolio with a short position in a bad news portfolio generates a 7.45% risk-adjusted return annually. This return anomaly appears on both news and non-news days. Overall, these findings suggest that the cross-sectional prediction of news is not fully incorporated into the stock price by investors.